requirements.txt
# Core dependencies
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
xgboost==1.7.6
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
joblib==1.3.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
jupyter==1.0.0

# API Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6